Update PoisonHelper.cs
Changing Mainhand Poison to "Instant Poison" and Offhand Poison to "Deadly Poison", usually it makes sense to have a fast weapon in the offhand, as the damage isn't that big, and the quicker the offhand, the more often the Deadly Poison stacks (up to 5) are applied. Additonally a slow Mainhand gives more Damage to Sinister Strike as it is calculated based on the Mainhands Damage, one of your main attacks in the Combat Rotation.